Rename `drivers/common` to `drivers/utils`
